+++ tardis/apps/idw/apps.py
@@ -1,0 +1,6 @@
+from tardis.app_config import AbstractTardisAppConfig
+
+
+class IDWConfig(AbstractTardisAppConfig):
+    name = "tardis.apps.idw"
+    verbose_name = "idw"
