+++ tardis/apps/idw/default_settings.py
@@ -1,0 +1,5 @@
+from tardis.default_settings import USER_MENU_MODIFIERS
+
+USER_MENU_MODIFIERS.extend(
+    ["tardis.apps.idw.user_menu_modifiers.add_idw_menu_item"]
+)
