+++ tardis/apps/idw/models/experiment.py
@@ -1,50 +1,52 @@
 from dataclasses import dataclass, field
 from datetime import datetime
-from pathlib import Path
 from typing import Any, Dict, List, Optional
 
 from yaml import SafeLoader
 
-from tardis.apps.yaml_dump.models.access_control import IAccessControl
-from tardis.apps.yaml_dump.models.dataclassification import IDataClassification
-from tardis.apps.yaml_dump.models.identifiers import IIdentifiers
-from tardis.apps.yaml_dump.models.yaml_dataclass import YAMLDataclass
+from tardis.apps.idw.models.access_control import IAccessControl
+from tardis.apps.idw.models.dataclassification import IDataClassification
+from tardis.apps.idw.models.identifiers import IIdentifiers
+from tardis.apps.idw.models.yaml_dataclass import YAMLDataclass
 
 
 @dataclass
-class Dataset(YAMLDataclass, IAccessControl, IDataClassification):
+class Experiment(YAMLDataclass, IAccessControl, IDataClassification):
     """
-    A class representing MyTardis Dataset objects.
+    A class representing MyTardis Experiment objects.
     """
 
-    yaml_tag = "!Dataset"
+    yaml_tag = "!Experiment"
     yaml_loader = SafeLoader
-    description: str = ""
+    title: str = ""
     data_status: int = 5
-    experiments: List[str] = field(default_factory=list)
-    instrument: str = ""
+    projects: List[str] = field(default_factory=list)
+    description: str = ""
     identifiers: list[str] = field(default_factory=list)
     metadata: Optional[Dict[str, Any]] = None
     object_schema: str = ""  # MTUrl in ingestion script
     # fields to add for updated data status
-    directory: Optional[Path] = None
-    immutable: bool = False
+    institution_name: Optional[str] = None
+    created_by: Optional[str] = None
+    url: Optional[str] = None  # MTUrl in ingestion script
+    locked: bool = False
+    start_time: Optional[datetime | str] = None
+    end_time: Optional[datetime | str] = None
     created_time: Optional[datetime | str] = None
-    modified_time: Optional[datetime | str] = None
+    update_time: Optional[datetime | str] = None
+    embargo_until: Optional[datetime | str] = None
     _store: Optional["IngestionMetadata"] = field(repr=False, default=None)
 
     def __post_init__(self) -> None:
-        """Dataclass lifecycle method that runs after an object is initialised.
-        This method initialises the identifier delegate class for this model."""
-        self.identifiers_delegate = DatasetIdentifiers(self)
+        self.identifiers_delegate = ExperimentIdentifiers(self)
 
 
-class DatasetIdentifiers(IIdentifiers):
-    """Dataset-specific methods related to identifiers."""
+class ExperimentIdentifiers(IIdentifiers):
+    """Experiment-specific methods related to identifiers."""
 
-    def __init__(self, dataset: Dataset):
-        self.dataset = dataset
-        super().__init__(dataset.identifiers)
+    def __init__(self, experiment: Experiment):
+        self.experiment = experiment
+        super().__init__(experiment.identifiers)
 
     def _is_unique(self, id: str) -> bool:
         """Private method to check whether an id is unique across all
@@ -56,10 +58,10 @@ class DatasetIdentifiers(IIdentifiers):
         Returns:
             bool: True if the identifier is unique, False if not.
         """
-        assert self.dataset._store is not None
+        assert self.experiment._store is not None
         return not any(
-            dataset.identifiers_delegate.has(id)
-            for dataset in self.dataset._store.datasets
+            experiment.identifiers_delegate.has(id)
+            for experiment in self.experiment._store.experiments
         )
 
     def add(self, value: str) -> bool:
@@ -76,8 +78,8 @@ class DatasetIdentifiers(IIdentifiers):
         return super().add(value) if self._is_unique(value) else False
 
     def update(self, old_id: str, id: str) -> bool:
-        """Updates an existing identifier in this Dataset and
-        all related Datafiles in the store. Checks if the identifier
+        """Updates an existing identifier in this Experiment and
+        all related Datasets in the store. Checks if the identifier
         is unique. Returns True if successful, False if not.
 
         Args:
@@ -87,18 +89,20 @@ class DatasetIdentifiers(IIdentifiers):
         Returns:
             bool: True if successfully updated, False if not unique.
         """
-        assert self.dataset._store is not None
+        assert self.experiment._store is not None
+        # Find all datasets and update their IDs.
         if not self._is_unique(id):
+            # Check if the new ID is unique.
             return False
-        # Find all experiments and update their IDs.
-        for datafile in self.dataset._store.datafiles:
-            if datafile.dataset == old_id:
-                datafile.dataset = id
+        for dataset in self.experiment._store.datasets:
+            if old_id in dataset.experiments:
+                dataset.experiments.remove(old_id)
+                dataset.experiments.append(id)
         return super().update(old_id, id)
 
     def delete(self, id_to_delete: str) -> bool:
-        """Deletes an identifier in this Dataset,
-        and updates identifiers in related Datafiles to use
+        """Deletes an identifier in this Experiment,
+        and updates identifiers in related Datasets to use
         an alternative identifier.
         Returns True if successfully deleted and updated, False if
         there are no other identifiers to use for related objects.
@@ -116,8 +120,9 @@ class DatasetIdentifiers(IIdentifiers):
             return False
         super().delete(id_to_delete)
         new_id = self.first()
-        assert self.dataset._store is not None
-        for datafile in self.dataset._store.datafiles:
-            if datafile.dataset == id_to_delete:
-                datafile.dataset = new_id
+        assert self.experiment._store is not None
+        for dataset in self.experiment._store.datasets:
+            if id_to_delete in dataset.experiments:
+                dataset.experiments.remove(id_to_delete)
+                dataset.experiments.append(new_id)
         return True
