+++ tardis/apps/idw/utils/dataset.py
@@ -5,8 +5,8 @@ from typing import Any, Dict
 
 from django.conf import settings
 
-from tardis.apps.yaml_dump.models.dataset import Dataset
-from tardis.apps.yaml_dump.utils.utility import (
+from tardis.apps.idw.models.dataset import Dataset
+from tardis.apps.idw.utils.utility import (
     add_acls_to_dataclass,
     add_data_classification_to_dataclass,
     add_metadata_to_dataclass,
