+++ tardis/apps/idw/utils/project.py
@@ -8,9 +8,9 @@ from typing import Any, Dict
 
 from django.conf import settings
 
-from tardis.apps.yaml_dump.models.project import Project
-from tardis.apps.yaml_dump.models.username import Username
-from tardis.apps.yaml_dump.utils.utility import (
+from tardis.apps.idw.models.project import Project
+from tardis.apps.idw.models.username import Username
+from tardis.apps.idw.utils.utility import (
     add_acls_to_dataclass,
     add_data_classification_to_dataclass,
     add_metadata_to_dataclass,
