--- tardis/apps/yaml_dump/apps.py
@@ -1,6 +1,0 @@
-from tardis.app_config import AbstractTardisAppConfig
-
-
-class YAMLDumpConfig(AbstractTardisAppConfig):
-    name = "tardis.apps.yaml_dump"
-    verbose_name = "yamldump"
