+++ tardis/urls/__init__.py
@@ -10,7 +10,7 @@ from django.views.static import serve
 
 from tardis.app_config import format_app_name_for_url, get_tardis_apps
 from tardis.apps.search.urls import urlpatterns as search_urls
-from tardis.apps.yaml_dump.urls import urlpatterns as yamldump_urls
+from tardis.apps.idw.urls import urlpatterns as idw_urls
 from tardis.tardis_portal.views import IndexView, login, rcauth, upload
 from tardis.tardis_portal.views.pages import site_routed_view
 
@@ -58,7 +58,7 @@ overridable_urls = [
 app_urls = []
 for app_name, app in get_tardis_apps():
     try:
-        if app_name in ["projects", "search", "yamldump"]:
+        if app_name in ["projects", "search", "idw"]:
             continue
         app_urls += [
             re_path(
@@ -74,7 +74,7 @@ urlpatterns = [
     re_path(r"^api/", include(api_urls)),
     # tastypie_swagger endpoints for API auto-documentation
     # re_path(r"^api/", include(tastypie_swagger_urls)),
-    re_path(r"^yaml/", include(yamldump_urls)),
+    re_path(r"^idw/", include(idw_urls)),
     # Experiment Views
     re_path(r"^experiment/", include(experiment_urls)),
     # Dataset Views
